Todos warnings OK

+++ src/proxy-server/utils.c
@@ -48,7 +48,7 @@ ssize_t write_until_enter_to_fd(int fd, buffer_t buffer) {
 
     ptr = buffer_read_ptr(buffer, &len);
     len--;
-    while( ptr[ret] != '\n' && ret < len) {
+    while( ptr[ret] != '\n' && (size_t)ret < len) {
         ret++;
     }
     ret++;
